Show weatcher feature; Change location feature.

+++ app/src/main/java/com/xsmile2008/righttests/entities/MainWeatherInfo.kt
@@ -4,8 +4,8 @@ import com.squareup.moshi.Json
 
 data class MainWeatherInfo(
         val temp: Double,
-        val humidity: Int,
-        val pressure: Int,
+        val humidity: Double,
+        val pressure: Double,
         @Json(name = "temp_min") val tempMin: Double,
         @Json(name = "temp_max") val tempMax: Double
 )
